Consolidate 7 milestone test files into test_invariants.py (#102)
Replace test_invariants_m1-m7.py with a single test_invariants.py
organized by topic (registry, decomposition, degree-1/2/3 invariants,
rotational/reflection behavior, public API, end-to-end). Deduplicates
mesh generators and _transform_tensors helper that were copied 5-7
times. Drops ~23 redundant tests (private-function rotation invariance
tests superseded by the public-API end-to-end suite; duplicate label
uniqueness and determinism tests already covered by TestPublicAPI).
133 tests remain (vs 156 across 7 files).
