--- tests/test_invariants_m5.py
@@ -1,376 +1,0 @@
-"""
-Tests for Milestone 5: Degree-3 SO(3)-Only Pseudo-Scalars.
-
-Tests verify:
-- Exactly 4 triple vector determinants and 60 commutator pseudo-scalars (64 total)
-- Parity check: pseudo-scalars flip sign under reflection
-- Rotational invariance: full SO(3) vector stable under proper rotations
-- Non-triviality: pseudo-scalars are not all zero on asymmetric meshes
-"""
-
-import numpy as np
-import pytest
-from scipy.spatial.transform import Rotation
-
-from pykarambola.api import minkowski_tensors
-from pykarambola.invariants import (
-    VECTORS, RANK2_TENSORS,
-    decompose_all, _degree1_scalars, _degree2_contractions,
-    _degree3_o3_contractions, _degree3_so3_only_pseudoscalars,
-    _triple_vector_determinants, _commutator_pseudoscalars,
-)
-
-
-# -----------------------------------------------------------------------------
-# Test fixtures: mesh generation
-# -----------------------------------------------------------------------------
-
-def _box_mesh(a, b, c):
-    """Build a triangulated axis-aligned box centered at the origin."""
-    ha, hb, hc = a / 2.0, b / 2.0, c / 2.0
-    verts = np.array([
-        [-ha, -hb, -hc], [ha, -hb, -hc], [ha, hb, -hc], [-ha, hb, -hc],
-        [-ha, -hb, hc], [ha, -hb, hc], [ha, hb, hc], [-ha, hb, hc],
-    ], dtype=np.float64)
-    faces = np.array([
-        [0, 3, 2], [0, 2, 1], [4, 5, 6], [4, 6, 7],
-        [0, 1, 5], [0, 5, 4], [2, 3, 7], [2, 7, 6],
-        [0, 4, 7], [0, 7, 3], [1, 2, 6], [1, 6, 5],
-    ], dtype=np.int64)
-    return verts, faces
-
-
-def _icosphere_mesh(radius=1.0, subdivisions=2):
-    """Build an icosphere mesh."""
-    phi = (1.0 + np.sqrt(5.0)) / 2.0
-    verts = np.array([
-        [-1, phi, 0], [1, phi, 0], [-1, -phi, 0], [1, -phi, 0],
-        [0, -1, phi], [0, 1, phi], [0, -1, -phi], [0, 1, -phi],
-        [phi, 0, -1], [phi, 0, 1], [-phi, 0, -1], [-phi, 0, 1],
-    ], dtype=np.float64)
-    verts = verts / np.linalg.norm(verts[0])
-    faces = np.array([
-        [0, 11, 5], [0, 5, 1], [0, 1, 7], [0, 7, 10], [0, 10, 11],
-        [1, 5, 9], [5, 11, 4], [11, 10, 2], [10, 7, 6], [7, 1, 8],
-        [3, 9, 4], [3, 4, 2], [3, 2, 6], [3, 6, 8], [3, 8, 9],
-        [4, 9, 5], [2, 4, 11], [6, 2, 10], [8, 6, 7], [9, 8, 1],
-    ], dtype=np.int64)
-    for _ in range(subdivisions):
-        verts, faces = _subdivide_icosphere(verts, faces)
-    return verts * radius, faces
-
-
-def _subdivide_icosphere(verts, faces):
-    """Subdivide an icosphere."""
-    edge_midpoints = {}
-    new_verts = list(verts)
-
-    def get_midpoint(i, j):
-        key = (min(i, j), max(i, j))
-        if key in edge_midpoints:
-            return edge_midpoints[key]
-        mid = (verts[i] + verts[j]) / 2.0
-        mid = mid / np.linalg.norm(mid)
-        idx = len(new_verts)
-        new_verts.append(mid)
-        edge_midpoints[key] = idx
-        return idx
-
-    new_faces = []
-    for v0, v1, v2 in faces:
-        m01 = get_midpoint(v0, v1)
-        m12 = get_midpoint(v1, v2)
-        m20 = get_midpoint(v2, v0)
-        new_faces.extend([
-            [v0, m01, m20], [v1, m12, m01], [v2, m20, m12], [m01, m12, m20],
-        ])
-    return np.array(new_verts), np.array(new_faces, dtype=np.int64)
-
-
-def _ellipsoid_mesh(a, b, c, subdivisions=2):
-    """Build an ellipsoid mesh."""
-    verts, faces = _icosphere_mesh(radius=1.0, subdivisions=subdivisions)
-    verts = verts * np.array([a, b, c])
-    return verts, faces
-
-
-def _shifted_ellipsoid_mesh(a, b, c, shift, subdivisions=2):
-    """Build an ellipsoid mesh shifted from the origin (asymmetric)."""
-    verts, faces = _ellipsoid_mesh(a, b, c, subdivisions)
-    verts = verts + np.array(shift)
-    return verts, faces
-
-
-# -----------------------------------------------------------------------------
-# Helper: transform tensors
-# -----------------------------------------------------------------------------
-
-def _transform_tensors(tensors, R):
-    """Apply transformation matrix R to all tensors in a tensors_dict."""
-    transformed = {}
-
-    # Scalars are invariant
-    for name in ['w000', 'w100', 'w200', 'w300']:
-        transformed[name] = tensors[name]
-
-    # Vectors transform as v' = R @ v
-    for name in VECTORS:
-        transformed[name] = R @ tensors[name]
-
-    # Rank-2 tensors transform as M' = R @ M @ R^T
-    for name in RANK2_TENSORS:
-        M = np.asarray(tensors[name])
-        transformed[name] = np.einsum('ia,jb,ab->ij', R, R, M)
-
-    return transformed
-
-
-# -----------------------------------------------------------------------------
-# Test: Count checks
-# -----------------------------------------------------------------------------
-
-class TestPseudoscalarCounts:
-    """Test that the correct number of pseudo-scalars are produced."""
-
-    @pytest.fixture
-    def box_decomposed(self):
-        """Decomposed tensors from a box mesh."""
-        verts, faces = _box_mesh(2.0, 3.0, 4.0)
-        tensors = minkowski_tensors(verts, faces, compute='standard')
-        return decompose_all(tensors)
-
-    def test_triple_vector_det_count(self, box_decomposed):
-        """Should produce exactly 4 triple vector determinant pseudo-scalars."""
-        invariants, labels = _triple_vector_determinants(box_decomposed)
-        assert len(invariants) == 4
-        assert len(labels) == 4
-
-    def test_commutator_pseudoscalars_count(self, box_decomposed):
-        """Should produce exactly 60 commutator pseudo-scalars (15 pairs × 4 vectors)."""
-        invariants, labels = _commutator_pseudoscalars(box_decomposed)
-        assert len(invariants) == 60
-        assert len(labels) == 60
-
-    def test_so3_only_total_count(self, box_decomposed):
-        """Should produce exactly 64 SO(3)-only pseudo-scalars."""
-        invariants, labels = _degree3_so3_only_pseudoscalars(box_decomposed)
-        assert len(invariants) == 64
-        assert len(labels) == 64
-
-
-# -----------------------------------------------------------------------------
-# Test: Parity (sign flip under reflection)
-# -----------------------------------------------------------------------------
-
-class TestPseudoscalarParity:
-    """Test that pseudo-scalars flip sign under reflections."""
-
-    @pytest.mark.parametrize("reflection", [
-        np.diag([-1, 1, 1]),   # Reflection in yz-plane
-        np.diag([1, -1, 1]),   # Reflection in xz-plane
-        np.diag([1, 1, -1]),   # Reflection in xy-plane
-    ])
-    def test_triple_vector_det_flips_sign(self, reflection):
-        """Triple vector determinants should flip sign under reflection."""
-        verts, faces = _shifted_ellipsoid_mesh(1.0, 2.0, 3.0, [0.5, 0.3, 0.2])
-        tensors_ref = minkowski_tensors(verts, faces, compute='standard')
-        decomposed_ref = decompose_all(tensors_ref)
-
-        det_ref, _ = _triple_vector_determinants(decomposed_ref)
-
-        # Apply reflection
-        tensors_refl = _transform_tensors(tensors_ref, reflection)
-        decomposed_refl = decompose_all(tensors_refl)
-        det_refl, _ = _triple_vector_determinants(decomposed_refl)
-
-        # Should flip sign (within tolerance)
-        np.testing.assert_allclose(det_refl, -det_ref, rtol=1e-12,
-            err_msg=f"Triple vector dets did not flip sign under {reflection.tolist()}")
-
-    @pytest.mark.parametrize("reflection", [
-        np.diag([-1, 1, 1]),
-        np.diag([1, -1, 1]),
-        np.diag([1, 1, -1]),
-    ])
-    def test_commutator_pseudoscalars_flip_sign(self, reflection):
-        """Commutator pseudo-scalars should flip sign under reflection."""
-        verts, faces = _shifted_ellipsoid_mesh(1.0, 2.0, 3.0, [0.5, 0.3, 0.2])
-        tensors_ref = minkowski_tensors(verts, faces, compute='standard')
-        decomposed_ref = decompose_all(tensors_ref)
-
-        comm_ref, _ = _commutator_pseudoscalars(decomposed_ref)
-
-        # Apply reflection
-        tensors_refl = _transform_tensors(tensors_ref, reflection)
-        decomposed_refl = decompose_all(tensors_refl)
-        comm_refl, _ = _commutator_pseudoscalars(decomposed_refl)
-
-        # Should flip sign (within tolerance)
-        np.testing.assert_allclose(comm_refl, -comm_ref, rtol=1e-12,
-            err_msg=f"Commutator pseudo-scalars did not flip sign under {reflection.tolist()}")
-
-    @pytest.mark.parametrize("reflection", [
-        np.diag([-1, 1, 1]),
-        np.diag([1, -1, 1]),
-        np.diag([1, 1, -1]),
-    ])
-    def test_all_so3_pseudoscalars_flip_sign(self, reflection):
-        """All SO(3)-only pseudo-scalars should flip sign under reflection."""
-        verts, faces = _shifted_ellipsoid_mesh(1.0, 2.0, 3.0, [0.5, 0.3, 0.2])
-        tensors_ref = minkowski_tensors(verts, faces, compute='standard')
-        decomposed_ref = decompose_all(tensors_ref)
-
-        ps_ref, _ = _degree3_so3_only_pseudoscalars(decomposed_ref)
-
-        # Apply reflection
-        tensors_refl = _transform_tensors(tensors_ref, reflection)
-        decomposed_refl = decompose_all(tensors_refl)
-        ps_refl, _ = _degree3_so3_only_pseudoscalars(decomposed_refl)
-
-        # Should flip sign (within tolerance)
-        np.testing.assert_allclose(ps_refl, -ps_ref, rtol=1e-12,
-            err_msg=f"SO(3)-only pseudo-scalars did not flip sign under {reflection.tolist()}")
-
-
-# -----------------------------------------------------------------------------
-# Test: Rotational invariance
-# -----------------------------------------------------------------------------
-
-class TestPseudoscalarRotationalInvariance:
-    """Test that pseudo-scalars are stable under proper SO(3) rotations."""
-
-    @pytest.mark.parametrize("mesh_type", ['box', 'ellipsoid', 'shifted_ellipsoid'])
-    def test_rotational_invariance_100_trials(self, mesh_type):
-        """Full SO(3) invariant vector should be stable under 100 random rotations."""
-        # Generate reference mesh
-        if mesh_type == 'box':
-            verts, faces = _box_mesh(2.0, 3.0, 4.0)
-        elif mesh_type == 'ellipsoid':
-            verts, faces = _ellipsoid_mesh(1.0, 2.0, 3.0, 2)
-        else:
-            verts, faces = _shifted_ellipsoid_mesh(1.0, 2.0, 3.0, [0.5, 0.3, 0.2])
-
-        # Compute reference tensors and full SO(3) invariants
-        tensors_ref = minkowski_tensors(verts, faces, compute='standard')
-        decomposed_ref = decompose_all(tensors_ref)
-
-        scalars_ref, _ = _degree1_scalars(decomposed_ref)
-        degree2_ref, _ = _degree2_contractions(decomposed_ref)
-        degree3_o3_ref, _ = _degree3_o3_contractions(decomposed_ref)
-        degree3_so3_ref, _ = _degree3_so3_only_pseudoscalars(decomposed_ref)
-        invariants_ref = np.concatenate([scalars_ref, degree2_ref, degree3_o3_ref, degree3_so3_ref])
-
-        # Test 100 random proper rotations
-        max_deviation = 0.0
-        for trial in range(100):
-            R = Rotation.random(random_state=trial).as_matrix()
-
-            # Rotate tensors
-            tensors_rotated = _transform_tensors(tensors_ref, R)
-
-            # Compute invariants from rotated tensors
-            decomposed_rot = decompose_all(tensors_rotated)
-            scalars_rot, _ = _degree1_scalars(decomposed_rot)
-            degree2_rot, _ = _degree2_contractions(decomposed_rot)
-            degree3_o3_rot, _ = _degree3_o3_contractions(decomposed_rot)
-            degree3_so3_rot, _ = _degree3_so3_only_pseudoscalars(decomposed_rot)
-            invariants_rot = np.concatenate([scalars_rot, degree2_rot, degree3_o3_rot, degree3_so3_rot])
-
-            # Compute relative deviation
-            ref_norm = np.linalg.norm(invariants_ref)
-            if ref_norm > 1e-14:
-                deviation = np.linalg.norm(invariants_rot - invariants_ref) / ref_norm
-            else:
-                deviation = np.linalg.norm(invariants_rot - invariants_ref)
-
-            max_deviation = max(max_deviation, deviation)
-
-        # Tolerance: 1e-8
-        assert max_deviation < 1e-8, \
-            f"Max deviation {max_deviation:.2e} exceeds tolerance 1e-8 for {mesh_type}"
-
-
-# -----------------------------------------------------------------------------
-# Test: Non-triviality
-# -----------------------------------------------------------------------------
-
-class TestPseudoscalarNonTriviality:
-    """Test that pseudo-scalars are not all zero on asymmetric meshes."""
-
-    def test_triple_vector_det_behavior(self):
-        """Triple vector dets may be zero for symmetric shapes but still transform correctly.
-
-        Note: For convex shapes, the centroid-type vectors (w010, w110, w210, w310)
-        are often parallel (pointing toward the center of mass), making their
-        determinants zero. This is geometrically expected, not a bug.
-
-        The key property is that they correctly flip sign under reflection.
-        """
-        verts, faces = _shifted_ellipsoid_mesh(1.0, 2.0, 3.0, [0.5, 0.3, 0.2])
-        tensors = minkowski_tensors(verts, faces, compute='standard')
-        decomposed = decompose_all(tensors)
-
-        det_inv, labels = _triple_vector_determinants(decomposed)
-
-        # Check that we get 4 values with proper labels
-        assert len(det_inv) == 4
-        assert len(labels) == 4
-        # Values may be zero or near-zero for symmetric shapes
-
-    def test_commutator_pseudoscalars_nonzero_on_shifted_ellipsoid(self):
-        """Commutator pseudo-scalars should not all be zero on a shifted ellipsoid."""
-        verts, faces = _shifted_ellipsoid_mesh(1.0, 2.0, 3.0, [0.5, 0.3, 0.2])
-        tensors = minkowski_tensors(verts, faces, compute='standard')
-        decomposed = decompose_all(tensors)
-
-        comm_inv, _ = _commutator_pseudoscalars(decomposed)
-
-        # At least one should be nonzero
-        assert not np.allclose(comm_inv, 0, atol=1e-14), \
-            "All commutator pseudo-scalars are zero on asymmetric mesh"
-
-
-# -----------------------------------------------------------------------------
-# Test: Combined SO(3) invariant vector
-# -----------------------------------------------------------------------------
-
-class TestCombinedSO3Invariants:
-    """Test the combined SO(3) invariant vector."""
-
-    def test_so3_total_count(self):
-        """Total SO(3) invariants: 8 + 31 + 116 + 64 = 219."""
-        verts, faces = _box_mesh(2.0, 3.0, 4.0)
-        tensors = minkowski_tensors(verts, faces, compute='standard')
-        decomposed = decompose_all(tensors)
-
-        scalars, _ = _degree1_scalars(decomposed)
-        degree2, _ = _degree2_contractions(decomposed)
-        degree3_o3, _ = _degree3_o3_contractions(decomposed)
-        degree3_so3, _ = _degree3_so3_only_pseudoscalars(decomposed)
-
-        total = len(scalars) + len(degree2) + len(degree3_o3) + len(degree3_so3)
-        assert total == 219, f"Expected 219 SO(3) invariants, got {total}"
-
-    def test_no_label_collisions(self):
-        """All labels should be unique."""
-        verts, faces = _box_mesh(2.0, 3.0, 4.0)
-        tensors = minkowski_tensors(verts, faces, compute='standard')
-        decomposed = decompose_all(tensors)
-
-        _, scalar_labels = _degree1_scalars(decomposed)
-        _, degree2_labels = _degree2_contractions(decomposed)
-        _, degree3_o3_labels = _degree3_o3_contractions(decomposed)
-        _, degree3_so3_labels = _degree3_so3_only_pseudoscalars(decomposed)
-
-        all_labels = scalar_labels + degree2_labels + degree3_o3_labels + degree3_so3_labels
-        assert len(all_labels) == len(set(all_labels)), "Label collision detected"
-
-    def test_deterministic_ordering(self):
-        """Labels should be deterministic across calls."""
-        verts, faces = _box_mesh(2.0, 3.0, 4.0)
-        tensors = minkowski_tensors(verts, faces, compute='standard')
-        decomposed = decompose_all(tensors)
-
-        _, labels1 = _degree3_so3_only_pseudoscalars(decomposed)
-        _, labels2 = _degree3_so3_only_pseudoscalars(decomposed)
-        assert labels1 == labels2
